implementation droits

+++ src/main/java/com/example/bety/controller/PersonController.java
@@ -20,6 +20,8 @@ import static org.springframework.http.HttpStatus.NOT_FOUND;
 public class PersonController {
     PersonService personService;
 
+    Person person = null;
+
     public PersonController(PersonService personService) {
         this.personService = personService;
     }
@@ -43,6 +45,9 @@ public class PersonController {
     @GetMapping("/deletePerson/{id}")
     public ResponseEntity<Void> deletePerson(@PathVariable Long id) {
 
+        if (person==null || !"admin".equals(person.getName())) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
         final boolean bool = personService.deletePerson(id);
 
         if (!bool) {
@@ -53,6 +58,11 @@ public class PersonController {
 
     @GetMapping("/addPerson/{a}/{b}/{c}")
     public ResponseEntity<Person> addPerson(@PathVariable String a, @PathVariable String b, @PathVariable String c) {
+
+        if (person==null || !"admin".equals(person.getName())) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         List<Role> listRole = new ArrayList<>();
         final Person person = new Person(null, a, b, c, listRole);
         return new ResponseEntity<>(personService.addPerson(person), HttpStatus.OK);
@@ -60,6 +70,11 @@ public class PersonController {
 
     @GetMapping("/updatePerson/{a}/{b}/{c}/{d}")
     public ResponseEntity<Person> updatePerson(@PathVariable Long a, @PathVariable String b, @PathVariable String c, @PathVariable String d) {
+
+        if (person==null || !"admin".equals(person.getName())) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         List<Role> listRole = new ArrayList<>();
         final Person person = new Person(a, b, c, d, listRole);
         return new ResponseEntity<>(personService.updatePerson(person), HttpStatus.OK);
@@ -80,5 +95,38 @@ public class PersonController {
     public ResponseEntity<List<Person>> showAllPersonsSortedLite() {
         return new ResponseEntity<>(personService.showAllPersonsSortedLite(), HttpStatus.OK);
     }
+
+    @GetMapping("/login/{name}/{password}")
+    public ResponseEntity<Boolean> login(@PathVariable String name, @PathVariable String password) {
+
+        HttpStatus httpStatus = HttpStatus.UNAUTHORIZED;
+
+        if (person != null) {
+            return new ResponseEntity<>(false, httpStatus);
+        }
+
+        if (person == null && personService.login(name, password)) {
+            httpStatus = HttpStatus.OK;
+
+            if ("admin".equals(name)) {
+                person = new Person(null, "admin", "admin", "admin", null);
+            } else {
+                person = personService.getPersonByName(name);
+            }
+
+
+        }
+
+
+        return new ResponseEntity<>(personService.login(name, password), httpStatus);
+    }
+
+    @GetMapping("/logout")
+    public ResponseEntity<Boolean> logout() {
+        person = null;
+        return new ResponseEntity<>(true, HttpStatus.OK);
+    }
+
+
 }
